Switch bet/bankroll amounts to decimal, rewrite SettleProfit, add DealerPlay
Bet/Bankroll/baseBet/netProfit and IBettingStrategy.GetNextBet were all
int, which can't represent a blackjack's 3:2 payout on an odd-numbered
bet (e.g. $5 -> $7.50). Switched to decimal throughout.

SettleProfit rewritten as an ordered if/else chain (most specific case
first) instead of OR-combined categories - the category version needed
defensive exclusions between branches and still missed two cases: a
player blackjack pushing against a non-blackjack dealer 21, and a
dealer blackjack pushing against a player's non-blackjack 21. Both are
covered by dedicated regression tests now.

DealerPlay.Play(dealerHand, shoe) is the fixed dealer rule (hit soft
17, stand hard 17+) - a static method with no state of its own, same
'don't abstract what's never swapped' reasoning as everywhere else.

Adds SettleProfitTests and DealerPlayTests.

+++ BlackjackStreakSimulator.Engine/FlatBettingStrategy.cs
@@ -2,7 +2,7 @@
 
 public class FlatBettingStrategy : IBettingStrategy
 {
-    public (int Bet, bool StreakCompleted) GetNextBet(int baseBet, int streakCount)
+    public (decimal Bet, bool StreakCompleted) GetNextBet(decimal baseBet, int streakCount)
     {
         return (baseBet, false); // never changes for flat betting
     }
